Major cleanup on the code
